+++ src/lib/ndpi_main.c
@@ -124,9 +124,6 @@ u_int ndpi_verbose_dga_detection = 0;
 static void *(*_ndpi_flow_malloc)(size_t size);
 static void (*_ndpi_flow_free)(void *ptr);
 
-static void *(*_ndpi_malloc)(size_t size);
-static void (*_ndpi_free)(void *ptr);
-
 /* ****************************************** */
 
 static ndpi_risk_info ndpi_known_risks[] = {
@@ -228,51 +225,8 @@ static inline uint8_t flow_is_proto(struct ndpi_flow_struct *flow, u_int16_t p) 
 
 /* ****************************************** */
 
-static volatile long int ndpi_tot_allocated_memory;
-
-/* ****************************************** */
-
-u_int32_t ndpi_get_tot_allocated_memory() {
-  return(__sync_fetch_and_add(&ndpi_tot_allocated_memory, 0));
-}
-
-/* ****************************************** */
-
-void *ndpi_malloc(size_t size) {
-  __sync_fetch_and_add(&ndpi_tot_allocated_memory, size);
-  return(_ndpi_malloc ? _ndpi_malloc(size) : malloc(size));
-}
-
-/* ****************************************** */
-
 void *ndpi_flow_malloc(size_t size) {
   return(_ndpi_flow_malloc ? _ndpi_flow_malloc(size) : ndpi_malloc(size));
-}
-
-/* ****************************************** */
-
-void *ndpi_calloc(unsigned long count, size_t size) {
-  size_t len = count * size;
-  void *p = ndpi_malloc(len);
-
-  if(p) {
-    memset(p, 0, len);
-    __sync_fetch_and_add(&ndpi_tot_allocated_memory, size);
-  }
-
-  return(p);
-}
-
-/* ****************************************** */
-
-void ndpi_free(void *ptr) {
-  if(_ndpi_free) {
-    if(ptr)
-      _ndpi_free(ptr);
-  } else {
-    if(ptr)
-      free(ptr);
-  }
 }
 
 /* ****************************************** */
@@ -285,39 +239,6 @@ void ndpi_flow_free(void *ptr) {
 }
 
 /* ****************************************** */
-
-void *ndpi_realloc(void *ptr, size_t old_size, size_t new_size) {
-  void *ret = ndpi_malloc(new_size);
-
-  if(!ret)
-    return(ret);
-  else {
-    if(ptr != NULL) {
-      memcpy(ret, ptr, (old_size < new_size ? old_size : new_size));
-      ndpi_free(ptr);
-    }
-    return(ret);
-  }
-}
-/* ****************************************** */
-
-char *ndpi_strdup(const char *s) {
-  if(s == NULL ){
-    return NULL;
-  }
-
-  int len = strlen(s);
-  char *m = ndpi_malloc(len + 1);
-
-  if(m) {
-    memcpy(m, s, len);
-    m[len] = '\0';
-  }
-
-  return(m);
-}
-
-/* *********************************************************************************** */
 
 /* Opaque structure defined here */
 struct ndpi_ptree
@@ -2515,16 +2436,10 @@ static int ndpi_add_host_ip_subprotocol(struct ndpi_detection_module_struct *ndp
   return(0);
 }
 
-void set_ndpi_malloc(void *(*__ndpi_malloc)(size_t size)) {
-  _ndpi_malloc = __ndpi_malloc;
-}
 void set_ndpi_flow_malloc(void *(*__ndpi_flow_malloc)(size_t size)) {
   _ndpi_flow_malloc = __ndpi_flow_malloc;
 }
 
-void set_ndpi_free(void (*__ndpi_free)(void *ptr)) {
-  _ndpi_free = __ndpi_free;
-}
 void set_ndpi_flow_free(void (*__ndpi_flow_free)(void *ptr)) {
   _ndpi_flow_free = __ndpi_flow_free;
 }
